+++ internal/handler/middleware/http/requestvalidation/handler.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package requestoverrides
+package requestvalidation
 
 import (
 	"net/http"
